requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
